refactor: flat 한 디렉터리 구조 변경
cardNumber, cardBrand, expiration, ownerName에 따라 디렉터리 분리
