+++ src/components/payments/index.ts
@@ -1,4 +1,4 @@
-export { default as CardNumberTextField } from './CardNumberTextField/CardNumberTextField';
-export { default as ExpirationDateTextField } from './ExpirationDateTextField/ExpirationDateTextField';
-export { default as OwnerNameTextField } from './OwnerNameTextField/OwnerNameTextField';
+export { default as CardNumberTextField } from './@cardNumber/CardNumberTextField/CardNumberTextField';
+export { default as ExpirationDateTextField } from './@expiration/ExpirationDateTextField/ExpirationDateTextField';
+export { default as OwnerNameTextField } from './@ownerName/OwnerNameTextField/OwnerNameTextField';
 export { default as PreviewCreditCard } from './PreviewCreditCard/PreviewCreditCard';
